+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 - Add validation for collection's spatial intervals
 - Add validation for collection's time intervals
+- Move `validate_bbox`, `validate_datetime` and `str_to_datetimes` functions from `api.search` to `shared` sub-module
+- Improve `bbox` validation for Antimeridian crossing bbox
 
 ## 3.2.0 (2025-03-20)